separated pauli logic from paulis.cpp
since paulis.cpp is an "API" file while the pauli logic previously therein was used by numerous core files.

A final inelegance remains; some unit test utilities leverage the internal pauli logic functions using extern, which will break when we eventually switch to private namespacing. Alas!
